Refactor kyt lib and bin
