+++ rs/bitcoin/kyt/src/fetch.rs
@@ -2,44 +2,34 @@ use crate::state::{
     FetchGuardError, FetchTxStatus, FetchTxStatusError, FetchedTx, HttpGetTxError,
     TransactionKytData,
 };
-use crate::types::{
-    CheckTransactionIrrecoverableError, CheckTransactionResponse, CheckTransactionRetriable,
-    CheckTransactionStatus,
-};
 use crate::{blocklist_contains, providers, state, Config};
 use bitcoin::Transaction;
 use futures::future::try_join_all;
 use ic_btc_interface::Txid;
+use ic_btc_kyt::{
+    get_tx_cycle_cost, CheckTransactionIrrecoverableError, CheckTransactionResponse,
+    CheckTransactionRetriable, CheckTransactionStatus, INITIAL_MAX_RESPONSE_BYTES,
+    RETRY_MAX_RESPONSE_BYTES,
+};
 use std::convert::Infallible;
 
 #[cfg(test)]
 mod tests;
 
-pub fn get_tx_cycle_cost(max_response_bytes: u32) -> u128 {
-    // 1 KiB for request, max_response_bytes for response
-    49_140_000 + 1024 * 5_200 + 10_400 * (max_response_bytes as u128)
+impl HttpGetTxError {
+    pub(crate) fn to_response(self, txid: Txid) -> CheckTransactionResponse {
+        let txid = txid.as_ref().to_vec();
+        match self {
+            HttpGetTxError::Rejected { message, .. } => {
+                CheckTransactionRetriable::TransientInternalError(message).into()
+            }
+            HttpGetTxError::ResponseTooLarge => {
+                (CheckTransactionIrrecoverableError::ResponseTooLarge { txid }).into()
+            }
+            _ => CheckTransactionRetriable::TransientInternalError(format!("{:?}", self)).into(),
+        }
+    }
 }
-
-/// Caller of check_transaction must attach this amount of cycles with the call.
-pub const CHECK_TRANSACTION_CYCLES_REQUIRED: u128 = 40_000_000_000;
-
-/// One-time charge for every check_transaction call.
-pub const CHECK_TRANSACTION_CYCLES_SERVICE_FEE: u128 = 100_000_000;
-
-// The max_response_bytes is initially set to 4kB, and then
-// increased to 400kB if the initial size isn't enough.
-// - The maximum size of a standard non-taproot transaction is 400k vBytes.
-// - Taproot transactions could be as big as full block size (4MiB).
-// - Currently a subnet's maximum response size is only 2MiB.
-// - Transaction size between 400kB and 2MiB are also uncommon, we could
-//   handle them in the future if required.
-// - Transactions bigger than 2MiB are very rare, and we can't handle them.
-
-/// Initial max response bytes is 4kB
-pub const INITIAL_MAX_RESPONSE_BYTES: u32 = 4 * 1024;
-
-/// Retry max response bytes is 400kB
-pub const RETRY_MAX_RESPONSE_BYTES: u32 = 400 * 1024;
 
 pub enum FetchResult {
     RetryWithBiggerBuffer,
@@ -273,7 +263,7 @@ pub trait FetchEnv {
                         );
                     }
                 }
-                FetchResult::Error(err) => error = Some((input_txid, err).into()),
+                FetchResult::Error(err) => error = Some(err.to_response(input_txid)),
                 FetchResult::RetryWithBiggerBuffer => (),
             }
         }
